% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{make_project}
\alias{make_project}
\title{Create ZEW project}
\usage{
make_project(path, type = c("Project report", "Paper"))
}
\arguments{
\item{path}{Path automatically set by research_project.dcf (see
\code{./rstudio/templates/project/})}

\item{type}{Choose between project report, slides and paper.}
}
\value{
Returns nothing.  See description above.
}
\description{
This function creates an R project which includes markdown or
quarto files as well as additional supporting files. These files create a
project in ZEW style. Important information on the usage can be found in the
README of the project as well as in the documents.
}
\details{
This function is employed when
a user selects ZEW Project.
}
\examples{
\dontrun{
  make_project(path = "~/test_project", type = "Quarto (.qmd)")
}
\dontrun{
  make_project(path = "~/test_project", type = "R Markdown (.Rmd)")
}
}
