% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_project.R
\name{new_project}
\alias{new_project}
\title{Create zewwww-econ research project}
\usage{
new_project(path)
}
\arguments{
\item{path}{Path automatically supplied by new_project.dcf when the user
chooses a directory in the RStudio New Project wizard.}
}
\value{
Invisibly returns \code{NULL}. Side-effects only.
}
\description{
Creates a full setup for research project with the standard folder
structure and Quarto templates for a Paper and Slides.
}
\details{
Called behind the scenes by the RStudio New Project wizard entry
  defined in \code{inst/rstudio/templates/project/new_project.dcf}.
}
