% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zew_colorstyle.R
\name{scale_color_zew}
\alias{scale_color_zew}
\title{Color scale constructor for ZEW colors}
\usage{
scale_color_zew(palette = "two", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of palette (see package README for comprehensive information).}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not, default is TRUE.}

\item{reverse}{Boolean indicating whether the palette should be reversed, default is FALSE.}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradient(), used respectively when discrete is TRUE or FALSE.}
}
\description{
This function implements ZEW corporate colors, which creates a fancy and uniform look.
}
\keyword{scale_color_zew}
