% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zew_plotstyle.R
\name{theme_zew}
\alias{theme_zew}
\title{Use ZEW style for plots}
\usage{
theme_zew(font = "", font_size = "12", flipped = FALSE, map = FALSE)
}
\arguments{
\item{font}{Select custom font, has to be loaded first using other functions.}

\item{font_size}{Select font size, default is 12.}

\item{flipped}{Boolean indicating whether the axes should be flipped, default is FALSE.}

\item{map}{Boolean indicating whether a map is plotted, default is FALSE.}
}
\description{
This function implements a ggplot2 theme, which creates a fancy and uniform look.
}
\examples{
line <- ggplot(line_df, aes(x = year, y = lifeExp)) +
geom_line(colour = "#007f7f", size = 1) +
geom_hline(yintercept = 0, size = 1, colour="#333333") +
theme_zew()
}
\keyword{theme_zew}
