% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zew_colorstyle.R
\name{zew_cols}
\alias{zew_cols}
\title{Function to extract ZEW colors as hex codes}
\usage{
zew_cols(...)
}
\arguments{
\item{...}{Character names of zew_colors}
}
\description{
Function to extract ZEW colors as hex codes
}
