% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zew_colorstyle.R
\name{zew_pal}
\alias{zew_pal}
\title{Return function to interpolate a zew color palette}
\usage{
zew_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in zew_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a zew color palette
}
