% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zewwwwEcontemplates-package.R
\docType{package}
\name{zewwwwEcontemplates-package}
\alias{zewwwwEcontemplates}
\alias{zewwwwEcontemplates-package}
\title{zewwwwEcontemplates: Create papers and presentations with 'Quarto'}
\description{
This package includes templates for papers and presentations in a nice layout.
}
\author{
\strong{Maintainer}: Simon Reif \email{simon.reif@zew.de}

Authors:
\itemize{
  \item Benedikt Stelter \email{benedikt.stelter@outlook.com}
}

}
\keyword{internal}
