% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_zewwww_project.R
\name{create_zewwww_project}
\alias{create_zewwww_project}
\title{Create zewwww project}
\usage{
create_zewwww_project(path, check_example)
}
\arguments{
\item{path}{Path automatically set by zewwww_project.dcf}

\item{check_example}{Choose whether to have examples or not}
}
\value{
Returns nothing.  See description above.
}
\description{
This function creates an R project which includes a quarto 
template as well as additional supporting files and a folder structure.
}
\details{
This function is employed behind the scenes when
a user creates a new zewwww project.
}
